Adds explicit file permissions
This came about due to issues in the environment from which I work in.
We are bound by strict CIS rules which caused the files dropped by
ansible to have a mode of 0600 and owned by user root.  Preventing the
`dd-agent` user from reading his own config files.

This PR simply enforces that user to own his the files such that he can
read them allowing the daemon to start.

Note that this PR assumes the user is dd-agent.  While this is true for
EL, I'm unaware of what the user is for Debian packages.