Trust new APT and RPM keys (#30)
Make the pkg managers trust the new keys in addition to the current keys.

* apt: Allow customizing the URL with `datadog_apt_key_url_new` (for users
with local apt mirrors).

* rpm: Download the new key through plain HTTP for best compatibility (since
SSL only works when the managed nodes have a recent version of python).
Ensure the integrity of the downloaded file using a sha-256 sum. We use
the `sha256sum` option instead of `checksum` since `checksum` was only
added in Ansible 2.0. The `rpm_key` module has a good idempotent behavior
so no need to have extra steps to check whether the key is already imported.