[readme] Add quotes on `apm_enabled` value
The only "truthy" value the trace agent currently understands
is "true". A true boolean in the playbook yaml is printed as "True" in
the generated datadog.conf, so the value has to be a string explicitly.

Other conf values don't have this issue since `dd-agent` does a pretty
good job at normalizing values from datadog.conf.

[skip ci]