Remove old INI config files from v6/v7 configuration (#272)
The conifiguration for these is now made in the main configuration file,
so these are no longer used. From the documentation for process-agent:
https://docs.datadoghq.com/infrastructure/process/?tab=linuxwindows
and tracing agent:
https://docs.datadoghq.com/tracing/send_traces/#configure-your-environment

You can see that now both reference the main configuration file.
You can also view the logs of the running agents to see that they're reading
the main agent config.