Add datadog_manage_config to disable changing the Agent config files (#410)
Adds a new variable datadog_manage_config that defaults to true but can be set to false. 
Doing so will make our role not change any of the Agent config files when it runs. 
This allows using the role to update the Agent if manual modifications have been made
to the config files without overwriting them.