Properly set agent_win_install_args in all cases (#497)
Initializes agent_win_install_args to win_install_args in all cases.
Removes useless line that sets agent_win_install_args to itself.

agent_win_install_args is used in pkg-windows.yml and pkg-windows-opts.yml. It is supposed to be set in pkg-windows-opts.yml, but all tasks that set it do it only under certain conditions.

On a default install, the role crashes.