unconditionally include linux install tasks
These are also responsible for setting up the installer, and doing some
bookkeeping around our GPG keys.
The installation of the agent package itself is already gated behind
agent_datadog_skip_install, so we won't attemp to reinstall the agent if
it's already installed to the latest version.
However we will now correctly probe for the installer ownership.
This should fix #603