[ABLD-27] Prepare deployment of Agent v7.70+ for macOS (arm64/x86_64) (#664)
* Prepare deployment of Agent v7.70+ for macOS (arm64/x86_64)

This change is meant to take into account how agent deliverables for
macOS embed architecture qualifiers (name & metadata) starting from
version 7.70 and some of the practical implications.

A word of caution: I'm far from being fluent with `ansible`, let alone
with the local setup and conventions, so the change is unlikely to work
out of the box nor to be exhaustive.

Following upstream changes might be of interest to reviewers:
- architecture qualifiers:
  - DataDog/omnibus-ruby#240
  - DataDog/agent-release-management#364
- new build for AArch64/ARM64:
  - DataDog/datadog-agent#37676
  - DataDog/agent-release-management#365

* Fix CI-reported failures

Altogether:
- use method `GET` w/ `return_content: false` instead of `HEAD` (not
  necessarily implemented)
- move `uri`-unsupported `retries` to upper level
- pass `status_code: 200` to populate
  `datadog_macos_agent7_latest_arch_check.status`
- make `ansible-lint` happy (at the expense of readability)

* Apply Fanny's suggestion

To prevent an undefined variable error.

Co-authored-by: Fanny Jiang <fanny.jiang@datadoghq.com>

* Make `ansible-lint` happy again

Use a folded block scalar to overcome the 160-char line limitation.

---------

Co-authored-by: Fanny Jiang <fanny.jiang@datadoghq.com>